Add on chain collection size

+++ programs/nft-voter/src/lib.rs
@@ -48,7 +48,7 @@ pub mod nft_voter {
     pub fn configure_collection(
         ctx: Context<ConfigureCollection>,
         weight: u64,
-        size: u32,
+        size: u64,
     ) -> Result<()> {
         log_version();
         instructions::configure_collection(ctx, weight, size)
